requirements.txt
folium
pywebview
selenium
cefpython